requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
six==1.15.0
uuid==1.30
Werkzeug==1.0.1
flask-cors==3.0.8
matplotlib==3.3.0
geopy==2.0.0 
Flask-Mail==0.9.1
